Clean up includes and forward declarations a bit.
