+++ src/settings/settings_manager.cpp
@@ -2,12 +2,8 @@
 
 #include <gflags/gflags.h>
 
-#include <memory>
-#include <string>
-#include <utility>
-#include <vector>
-
 #include "common/macros.h"
+#include "execution/sql/value_util.h"
 #include "main/db_main.h"
 #include "parser/expression/constant_value_expression.h"
 #include "settings/settings_callbacks.h"
