+++ test/test_util/tpch/workload.cpp
@@ -5,6 +5,7 @@
 #include "common/managed_pointer.h"
 #include "execution/exec/execution_context.h"
 #include "execution/execution_util.h"
+#include "execution/sql/value_util.h"
 #include "execution/table_generator/table_generator.h"
 #include "main/db_main.h"
